refs #148 Отдельный файл для реализации параллельной поразрядной сортировки
