+++ benchmark/burst/algorithm/radix_sort.cpp
@@ -10,6 +10,7 @@
 #include <iostream>
 #include <random>
 #include <string>
+#include <thread>
 #include <unordered_map>
 #include <vector>
